# src/V1.0.0/doc/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Task management API",
      "contact": {},
      "version": "v1"
    },
    "servers": [
      {
        "url": "http://localhost:3001"
      }
    ],
    "paths": {
...
      "/v1/projects/{id}/tasks": {
        "get": {
          "tags": ["Project"],
          "summary": "Get Tasks by projectId",
          "parameters": [
            {
              "name": "id",
              "description": "id of project",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Return a tasks",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Something went wrong",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                }
              }
            }
          }
        }
      },
      "/V1/task": {
        "post": {
          "tags": ["Task"],
          "summary": "Create Task",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/createTask"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/createTask"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/createTask"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Task created successfully",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Something went wrong",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                }
              }
            }
          }
        },
        "get": {
          "tags": ["Task"],
          "summary": "Get Task list",
          "responses": {
            "200": {
              "description": "Return a task",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Something went wrong",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                }
              }
            }
          }
        }
      },
      "/v1/task/{id}": {
        "get": {
          "tags": ["Task"],
          "summary": "Get Task by Id",
          "parameters": [
            {
              "name": "id",
              "description": "id of task",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Return a task",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Something went wrong",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": ["Task"],
          "summary": "Update Task by Id",
          "parameters": [
            {
              "name": "id",
              "description": "id of task",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/createTask"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/createTask"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/createTask"
                }
              }
            }
          },
          "responses": {
            "200": {
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Something went wrong",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": ["Task"],
          "summary": "Delete Task by Id",
          "parameters": [
            {
              "name": "id",
              "description": "id of task",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Something went wrong",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/errorResponse500"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
...
          "createTask": {
            "required": [
              "title",
              "description",
              "status",
              "projectId"
            ],
            "type": "object",
            "properties": {
              "title": {
                "type": "string",
                "description": "Title of task"
              },
              "description": {
                "type": "string",
                "description": "Description of task"
              },
              "status": {
                "type": "string",
                "description": "Status of task"
              },
              "projectId": {
                "type": "string",
                "description": "Id of project"
              }
            },
            "example": {
              "title": "user crud API",
              "description": "description",
              "status": "Not started",
              "projectId": "fcd2cf05-f96a-4a82-bb9c-1c8c0db5042a"
            }
          },
        "SuccessResponse": {
          "type": "object",
          "properties": {
            "isSuccess": {
              "type": "boolean",
              "example": true
            },
            "message": {
              "type": "string",
              "example": "message"
            },
            "code": {
              "type": "integer",
              "example": "200"
            },
            "data": {
              "type": "object"
            }
          },
          "additionalProperties": false
        },
...
        "errorResponse500": {
          "type": "object",
          "properties": {
            "isSuccess": {
              "type": "boolean",
              "example": false
            },
            "message": {
              "type": "string",
              "example": "Internal server error"
            },
            "code": {
              "type": "integer",
              "example": "500"
            },
            "data": {
              "type": "object"
            }
          },
          "additionalProperties": false
        }
      }
    },
...
  }